// src/api/openapi/auto-generated/models/ResultSetType.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter API
 * Welcome to Jupiter\'s ClimateScore Global API
 *
 * The version of the OpenAPI document: 2.2.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const ResultSetType = {
    perils: 'perils',
    scores: 'scores',
    mesh: 'mesh',
    damages: 'damages',
    portfolioFinancial: 'portfolio_financial',
    supplemental: 'supplemental',
    disclosure: 'disclosure',
    unknownDefaultOpenApi: '11184809'
} as const;
export type ResultSetType = typeof ResultSetType[keyof typeof ResultSetType];


export function instanceOfResultSetType(value: any): boolean {
    for (const key in ResultSetType) {
        if (Object.prototype.hasOwnProperty.call(ResultSetType, key)) {
            if ((ResultSetType as Record<string, ResultSetType>)[key] === value) {
                return true;
            }
        }
    }
    return false;
}

export function ResultSetTypeFromJSON(json: any): ResultSetType {
    return ResultSetTypeFromJSONTyped(json, false);
}

export function ResultSetTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ResultSetType {
    return json as ResultSetType;
}

export function ResultSetTypeToJSON(value?: ResultSetType | null): any {
    return value as any;
}